train-on-gcp/docker_requirements.txt
mlagents
cloudml-hypertune
